--- Content.Shared/Spider/IgnoreSpiderWebComponent.cs
@@ -1,7 +1,0 @@
-namespace Content.Shared.Spider;
-
-[RegisterComponent]
-public sealed partial class IgnoreSpiderWebComponent : Component
-{
-
-}
