--- Content.Shared/Spider/SpiderWebObjectComponent.cs
@@ -1,9 +1,0 @@
-using Robust.Shared.GameStates;
-
-namespace Content.Shared.Spider;
-
-[RegisterComponent, NetworkedComponent]
-[Access(typeof(SharedSpiderSystem))]
-public sealed partial class SpiderWebObjectComponent : Component
-{
-}
