--- Content.Shared/Spider/SpiderWebVisualsComponent.cs
@@ -1,9 +1,0 @@
-using Robust.Shared.Serialization;
-
-namespace Content.Shared.Spider;
-
-[Serializable, NetSerializable]
-public enum SpiderWebVisuals
-{
-    Variant
-}
